Feature: Add Vnet-only resource groups to graph

+++ AzViz/src/private/ConvertTo-DOTLangauge.ps1
@@ -218,7 +218,9 @@ function ConvertTo-DOTLanguage {
                 }
 
                 $NodesAndEdges = $nodes + $edges
+            }
 
+            if ($Resources -or $VNets) {
                 $ResourceGroupLocation = (Get-AzResourceGroup -Name $Target.Name -Verbose:$false).Location
                 $ResourceGroupSubGraphName = [string]::Concat($(Remove-SpecialChars -String $Target.Name -SpecialChars '() []{}&-'), $Counter)
                 $ResourceGroupSubGraphNameLabel = Get-ImageLabel -Type "ResourceGroups" -Row1 "ResourceGroup: $(Remove-SpecialChars -String $Target.name)" -Row2 "Location: $($ResourceGroupLocation)"
@@ -237,9 +239,7 @@ function ConvertTo-DOTLanguage {
                     $NetworkLayout
                     $NodesAndEdges
                 }
-
-            }
-            else {
+            } else {
                 Write-CustomHost -String "No resources found.. re-run the command and try increasing the category depth using -CategoryDepth 2 or -CategoryDepth 3 cmdlet parameters." -Indentation 1 -Color Red -AddTime
             }
             #endregion plotting-edges-to-nodes
